==> SeroScore/ACS/12-11-2022-Diff_Pro_vs_nonPro_threeInterval_risk5score_FcR/4_Define_SeroScore_LassoFeatureSelection_wRisk6.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r}
meta_colors <- list(
  "group" = c(
    "NA" = "gray", 
    "progressor" = "#AD07E3",
    "non-progressor" = "#238b45"
    )
)
``` 



```{r}
library(DMwR2)
#Data a folder to save the results.
output_dir <- 'Progressor_allmonths'

Data_group <- readRDS(file.path(output_dir, "Data_group.RDS"))
Data_group <- factor(Data_group, levels = c("non-progressor", "progressor"))
Data_combining <- readRDS(file.path(output_dir, "Data_combining.RDS"))

Data_combining <- scale(Data_combining, center = TRUE, scale = TRUE)
```

```{r}

library(dplyr)
folder <- './Classification_RF_LASSO'
data.df <- readRDS(file.path(folder, "Features_Freq.RDS"))

temp.df <- top_n(data.df, 19, Freq)
interested.features <- as.character(temp.df$FeatureName)
# interested.features <- c("Risk6", "IgG1_LAM", "IgG4_Ag85", "FcgR2A_PPD","FcgR3B_PPD",
#                          "IgA1_PPD", "IgG2_LAM", "IgG4_Hspx", "IgM_ESAT6", 
#                          "IgA2_Tbad", "IgG1_Hspx", "IgG1_PPD", "IgG3_CFP10")
interested.features <- gsub("\\.", " ", interested.features)
Data_combining_selected <- Data_combining[, interested.features]
y = Data_group
y = gsub("non-progressor", 0, y)
y = gsub("progressor", 1, y)

data.df <- data.frame(y, Data_combining_selected)
```


```{r}
#Create a folder to save the results.
saved.dir <- './SeroScore_FcR_Lasso_wRisk6'
if (!dir.exists(saved.dir)){
  dir.create((saved.dir))
}
```

Explore all the possible combination and evaluate them by ROC.
```{r}
FcR <- interested.features[1:19]

n <- length(FcR)
l <- rep(list(0:1), n)

FcR.index <- expand.grid(l)
FcR.index <- FcR.index[-1, ]
FcR.Data <- Data_combining_selected[, FcR]
colnames(FcR.Data) <- FcR
saveRDS(FcR.Data, file.path(saved.dir, "FcR.Data.RDS"))
```

```{r}
# Negative <- interested.features[13:16]
# 
# n <- length(Negative)
# l <- rep(list(0:1), n)
# 
# Neg.index <- expand.grid(l)
# Neg.index <- Neg.index[-1, ]
# Neg.Data <- Data_combining_selected[, Negative]
# colnames(Neg.Data) <- Negative
# saveRDS(Neg.Data, file.path(saved.dir, "Neg.Data.RDS"))
```

Calculate the seroScore.
```{r}
seroScore <- data.frame(matrix(NA, nrow = dim(Data_combining)[1], ncol = dim(FcR.index)[1]))
curr.index <- 1
for(index.p in 1 : dim(FcR.index)[1]){
    #for(index.n in 1 : dim(Neg.index)[1]){
      #For positive parts.
      temp.FcR.index <- FcR.index[index.p, ]
      temp.FcR.index <- which(temp.FcR.index == 1)
      if(length(temp.FcR.index) > 1){
        Temp.pos <- rowSums(FcR.Data[, temp.FcR.index]) #Sum each rows.
      }else{
        Temp.pos <- FcR.Data[, temp.FcR.index]
      }
    
    
      #Calculate the difference.
      Temp.value <- Temp.pos
      seroScore[, curr.index] <- Temp.value
      curr.index <- curr.index + 1
    #}
}

saveRDS(seroScore, file.path(saved.dir, 'SeroScore.RDS'))
write.csv(seroScore, file.path(saved.dir, "SeroScore.csv"))
```
